Added CZ comments to VRLifeAPI library

+++ src/VrLifeAPI/Common/Core/Services/ServiceUtils.cs
@@ -8,8 +8,19 @@ using VrLifeAPI.Networking.NetworkingModels;
 
 namespace VrLifeAPI.Common.Core.Services
 {
+    /// <summary>
+    /// Extension třída pro služby
+    /// </summary>
     public static class ServiceUtils
     {
+        /// <summary>
+        /// Vytvoření chybové zprávy na událost.
+        /// </summary>
+        /// <param name="msgId">ID chybné zprávy.</param>
+        /// <param name="errType">ID typu chyby.</param>
+        /// <param name="errCode">ID chybové hlášky.</param>
+        /// <param name="errMsg">Textová chybová hláška.</param>
+        /// <returns>EventResponse objekt k odeslání</returns>
         public static EventResponse CreateErrorResponse(ulong msgId, uint errType, uint errCode, string errMsg = null)
         {
             EventResponse msg = new EventResponse();
@@ -21,6 +32,14 @@ namespace VrLifeAPI.Common.Core.Services
             return msg;
         }
 
+        /// <summary>
+        /// Vytvoření chybové zprávy.
+        /// </summary>
+        /// <param name="msgId">ID chybné zprávy.</param>
+        /// <param name="errType">ID typu chyby.</param>
+        /// <param name="errCode">ID chybové hlášky.</param>
+        /// <param name="errMsg">Textová chybová hláška.</param>
+        /// <returns>MainMessage objekt k odeslání.</returns>
         public static MainMessage CreateErrorMessage(ulong msgId, uint errType, uint errCode, string errMsg = null)
         {
             ErrorMsg errorMsg = new ErrorMsg();
@@ -36,6 +55,11 @@ namespace VrLifeAPI.Common.Core.Services
 
         }
 
+        /// <summary>
+        /// Vytvoření stavové zprávy OK.
+        /// </summary>
+        /// <param name="msgId">Id přijaté zprávy.</param>
+        /// <returns>MainMessage objekt k odeslání.</returns>
         public static MainMessage CreateOkMessage(ulong msgId = 0)
         {
             OkMsg okMsg = new OkMsg();
@@ -47,6 +71,13 @@ namespace VrLifeAPI.Common.Core.Services
             return msg;
         }
 
+        /// <summary>
+        /// Vytvoření zprávy přesměrování.
+        /// </summary>
+        /// <param name="recvMsg">Přijatá zpráva.</param>
+        /// <param name="address">Adresa na kterou je zpráva přesměrována.</param>
+        /// <param name="port">Port na který je zpráva přesměrována.</param>
+        /// <returns>MainMessage objekt k odeslání.</returns>
         public static MainMessage CreateRedirectMessage(MainMessage recvMsg, int address, int port)
         {
             RedirectMsg redirectMsg = new RedirectMsg();
@@ -59,11 +90,22 @@ namespace VrLifeAPI.Common.Core.Services
             return msg;
         }
 
+        /// <summary>
+        /// Vytvoření zprávy přesměrování. 
+        /// </summary>
+        /// <param name="recvMsg">Přijatá zpráva.</param>
+        /// <param name="ip">Adresa na kterou je zpráva přesměrována</param>
+        /// <returns>MainMessage objekt k odeslání.</returns>
         public static MainMessage CreateRedirectMessage(MainMessage recvMsg, IPEndPoint ip)
         {
-            return ServiceUtils.CreateRedirectMessage(recvMsg, (int)ip.Address.ToInt(), ip.Port);
+            return ServiceUtils.CreateRedirectMessage(recvMsg, ip.Address.ToInt(), ip.Port);
         }
 
+        /// <summary>
+        /// Kontrola zda je zpráva chybovou systémovou hláškou.
+        /// </summary>
+        /// <param name="msg">Zpráva ke kontrole.</param>
+        /// <returns>true - chybová zpráva, false - jiná</returns>
         public static bool IsError(MainMessage msg)
         {
             return msg.MessageTypeCase == MainMessage.MessageTypeOneofCase.SystemMsg &&
